+++ src/DomHandler/Wsdl/Tag/AbstractTagOperationElement.php
@@ -36,7 +36,7 @@ abstract class AbstractTagOperationElement extends AbstractTag
         return $this->hasAttribute(self::ATTRIBUTE_MESSAGE) ? $this->getAttribute(self::ATTRIBUTE_MESSAGE)->getValueNamespace() : '';
     }
     /**
-     * @return TagMessage
+     * @return TagMessage|null
      */
     public function getMessage()
     {