+++ src/File/ClassMap.php
@@ -47,6 +47,7 @@ class ClassMap extends AbstractModelFile
     }
     /**
      * @param MethodContainer $methods
+     * @return ClassMap
      */
     protected function getClassMethods(MethodContainer $methods)
     {
@@ -119,6 +120,6 @@ class ClassMap extends AbstractModelFile
      */
     protected function getStructName(StructModel $struct)
     {
-        return substr($struct->getPackagedName(true), $struct->getNamespace() != '' ? 1 : 0);
+        return str_replace('\\', '\\\\', $struct->getPackagedName(true));
     }
 }