+++ src/File/Composer.php
@@ -22,7 +22,6 @@ class Composer extends AbstractFile
     {
         $composer = new Application();
         $composer->setAutoExit(false);
-
         $composer->run(new ArrayInput(array(
             'command' => 'init',
             '--verbose' => true,
@@ -32,13 +31,11 @@ class Composer extends AbstractFile
             '--require' => array(
                 'php:>=5.3.3',
                 'ext-soap:*',
-                'wsdltophp/packagebase:dev-master',
+                'wsdltophp/packagebase:~1.0',
             ),
             '--working-dir' => $this->getGenerator()->getOptionDestination(),
         )));
-
-        $this->addAutoloadToComposerJson();
-
+        $this->completeComposerJson();
         if ($this->getRunComposerUpdate() === true) {
             return $composer->run(new ArrayInput(array(
                 'command' => 'update',
@@ -53,15 +50,31 @@ class Composer extends AbstractFile
     /**
      * @return Composer
      */
-    protected function addAutoloadToComposerJson()
+    protected function completeComposerJson()
     {
         $content = $this->getComposerFileContent();
         if (is_array($content) && !empty($content)) {
-            $content['autoload'] = array(
-                'psr-4' => $this->getPsr4Autoload(),
-            );
+            $this->addAutoloadToComposerJson($content)->addComposerSettings($content);
         }
         return $this->setComposerFileContent($content);
+    }
+    /**
+     * @return Composer
+     */
+    protected function addAutoloadToComposerJson(array &$content)
+    {
+        $content['autoload'] = array(
+            'psr-4' => $this->getPsr4Autoload(),
+        );
+        return $this;
+    }
+    /**
+     * @return Composer
+     */
+    protected function addComposerSettings(array &$content)
+    {
+        $content = array_merge_recursive($content, $this->getGenerator()->getOptionComposerSettings());
+        return $this;
     }
     /**
      * @return array