+++ src/File/Validation/StringRule.php
@@ -13,12 +13,11 @@ class StringRule extends AbstractRule
      */
     public function applyRule($parameterName, $value, $itemType = false)
     {
-        $this
-            ->getMethod()
-                ->addChild('// validation for constraint: string')
-                ->addChild(sprintf('if (!is_null($%1$s) && !is_string($%1$s)) {', $parameterName))
-                    ->addChild($this->getMethod()->getIndentedString(sprintf('throw new \InvalidArgumentException(sprintf(\'Invalid value, please provide a string, "%%s" given\', gettype($%s)), __LINE__);', $parameterName), 1))
-                ->addChild('}');
+        $this->getMethod()
+            ->addChild('// validation for constraint: string')
+            ->addChild(sprintf('if (!is_null($%1$s) && !is_string($%1$s)) {', $parameterName))
+            ->addChild($this->getMethod()->getIndentedString(sprintf('throw new \InvalidArgumentException(sprintf(\'Invalid value, please provide a string, "%%s" given\', gettype($%s)), __LINE__);', $parameterName), 1))
+            ->addChild('}');
         return $this;
     }
 }