+++ src/Generator/Utils.php
@@ -74,14 +74,13 @@ class Utils
     public static function getContentFromUrl($url, $basicAuthLogin = null, $basicAuthPassword = null, $proxyHost = null, $proxyPort = null, $proxyLogin = null, $proxyPassword = null, array $contextOptions = array())
     {
         $context = null;
-        $options = self::getContentFromUrlContextOptions($url, $basicAuthLogin, $basicAuthPassword, $proxyHost, $proxyPort, $proxyLogin, $proxyPassword, $contextOptions);
+        $options = self::getStreamContextOptions($basicAuthLogin, $basicAuthPassword, $proxyHost, $proxyPort, $proxyLogin, $proxyPassword, $contextOptions);
         if (!empty($options)) {
             $context = stream_context_create($options);
         }
         return file_get_contents($url, false, $context);
     }
     /**
-     * @param string $url
      * @param string $basicAuthLogin
      * @param string $basicAuthPassword
      * @param string $proxyHost
@@ -91,13 +90,12 @@ class Utils
      * @param array $contextOptions
      * @return string[]
      */
-    public static function getContentFromUrlContextOptions($url, $basicAuthLogin = null, $basicAuthPassword = null, $proxyHost = null, $proxyPort = null, $proxyLogin = null, $proxyPassword = null, array $contextOptions = array())
+    public static function getStreamContextOptions($basicAuthLogin = null, $basicAuthPassword = null, $proxyHost = null, $proxyPort = null, $proxyLogin = null, $proxyPassword = null, array $contextOptions = array())
     {
-        $protocol = strpos($url, 'https://') !== false ? 'https' : 'http';
         $proxyOptions = $basicAuthOptions = array();
         if (!empty($basicAuthLogin) && !empty($basicAuthPassword)) {
             $basicAuthOptions = array(
-                $protocol => array(
+                'http' => array(
                     'header' => array(
                         sprintf('Authorization: Basic %s', base64_encode(sprintf('%s:%s', $basicAuthLogin, $basicAuthPassword))),
                     ),
@@ -106,11 +104,8 @@ class Utils
         }
         if (!empty($proxyHost)) {
             $proxyOptions = array(
-                $protocol => array(
-                    'proxy' => sprintf('tcp://%s%s',
-                        $proxyHost,
-                        empty($proxyPort) ? '' : sprintf(':%s', $proxyPort)
-                    ),
+                'http' => array(
+                    'proxy' => sprintf('tcp://%s%s', $proxyHost, empty($proxyPort) ? '' : sprintf(':%s', $proxyPort)),
                     'header' => array(
                         sprintf('Proxy-Authorization: Basic %s', base64_encode(sprintf('%s:%s', $proxyLogin, $proxyPassword))),
                     ),
@@ -122,15 +117,32 @@ class Utils
     /**
      * Returns the value with good type
      * @param mixed $value the value
+     * @param string $knownType the value
      * @return mixed
      */
     public static function getValueWithinItsType($value, $knownType = null)
     {
-        if (is_int($value) || (!is_null($value) && in_array($knownType, array('time', 'positiveInteger', 'unsignedLong', 'unsignedInt', 'short', 'long', 'int', 'integer'), true))) {
+        if (is_int($value) || (!is_null($value) && in_array($knownType, array(
+            'time',
+            'positiveInteger',
+            'unsignedLong',
+            'unsignedInt',
+            'short',
+            'long',
+            'int',
+            'integer',
+        ), true))) {
             return intval($value);
-        } elseif (is_float($value) || (!is_null($value) && in_array($knownType, array('float', 'double', 'decimal'), true))) {
+        } elseif (is_float($value) || (!is_null($value) && in_array($knownType, array(
+            'float',
+            'double',
+            'decimal',
+        ), true))) {
             return floatval($value);
-        } elseif (is_bool($value) || (!is_null($value) && in_array($knownType, array('bool', 'boolean'), true))) {
+        } elseif (is_bool($value) || (!is_null($value) && in_array($knownType, array(
+            'bool',
+            'boolean',
+        ), true))) {
             return ($value === 'true' || $value === true || $value === 1 || $value === '1');
         }
         return $value;
@@ -148,7 +160,6 @@ class Utils
                 $destination = substr($destination, 2);
             }
             $destinationParts = explode('/', $destination);
-
             $fileParts = pathinfo($origin);
             $fileBasename = (is_array($fileParts) && array_key_exists('basename', $fileParts)) ? $fileParts['basename'] : '';
             $parts = parse_url(str_replace('/' . $fileBasename, '', $origin));
@@ -158,7 +169,6 @@ class Utils
             $path = str_replace('/' . $fileBasename, '', $path);
             $pathParts = explode('/', $path);
             $finalPath = implode('/', $pathParts);
-
             foreach ($destinationParts as $locationPart) {
                 if ($locationPart == '..') {
                     $finalPath = substr($finalPath, 0, strrpos($finalPath, '/', 0));
@@ -166,7 +176,6 @@ class Utils
                     $finalPath .= '/' . $locationPart;
                 }
             }
-
             $port = (is_array($parts) && array_key_exists('port', $parts)) ? $parts['port'] : '';
             /**
              * Remote file
@@ -223,7 +232,7 @@ class Utils
     public static function removeNamespace($namespacedClassName)
     {
         $elements = explode('\\', $namespacedClassName);
-        return (string)array_pop($elements);
+        return (string) array_pop($elements);
     }
     /**
      * @param string $directory